close #194 -Postmapping

+++ backend/src/main/java/com/sutse/team06/entity/TransportationCar.java
@@ -45,7 +45,7 @@ public class TransportationCar{
 	   private Employee empId;
 	   
 	   private @NonNull String transportDetial;
-	   private @NonNull int amountPeople;
+	   private @NonNull Long amountPeople;
 
 	
 }
